pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geoai"
version = "0.2.0"
description = "Open geolocation training stack with hierarchical prediction and A100-ready workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "huggingface-hub>=0.30",
  "numpy",
  "Pillow",
  "pydantic>=2",
  "pyyaml",
  "requests",
  "rich",
  "torch",
  "torchvision",
  "timm",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
